assetmanager/ui/requirements.txt
google-cloud-aiplatform>=1.44.0
google-api-python-client>=2.122.0
google-cloud-core>=2.4.1
graphdatascience
retry==0.9.2
langchain>=0.1.11
langchain-google-vertexai>=0.1.1
langchain-community>=0.0.27
streamlit>=1.27.2
streamlit-chat>=0.1.1
streamlit-chat-media>=0.0.4
numpy
pandas
requests
plotly==5.15.0
tqdm
pillow
altair<5